update .gitignore to ignore clangd and pyenv files
By default, `clangd`-a C/C++ language server-creates a `.cache/`
directory in the same directory as the `compile_commands.json` file.
Additionally, `pyenv`-a Python version and environment manager-creates a
`.python-version` file in a directory to tell the user's shell which
Python environment to use. This commit ignores both these files.